FIX - Travis tests - replaced `python setup.py install` with `pip install -e .` in `.travis.yml`; - changed file utility version call param; and - moved database initialisation from `database.py` to `model.py`.
