+++ corptest/database.py
@@ -22,10 +22,3 @@ DB_SESSION = scoped_session(sessionmaker(autocommit=False,
                                          bind=ENGINE))
 BASE = declarative_base()
 BASE.query = DB_SESSION.query_property()
-
-def init_db():
-    """Initialise the database."""
-    # import all modules here that might define models so that
-    # they will be registered properly on the metadata.  Otherwise
-    # you will have to import them first before calling init_db()
-    BASE.metadata.create_all(bind=ENGINE)
